refac: select kind component
