+++ components/definitions/NewEntityModal.tsx
@@ -1,15 +1,8 @@
-import { Portal, ToggleButton, Modal, TextInput,  Text, Button } from "react-native-paper"
-import { View } from "react-native"
+import { Portal, Modal, TextInput,  Text, Button } from "react-native-paper"
 import { Kind, fromJSON, Module } from "wollok-ts/dist/model"
 import React, { useState } from "react"
 import { useTheme } from "@react-navigation/native"
-
-const entityKinds: {kind: Kind, description: string}[] = 
-    [
-        {kind:'Singleton', description: 'OBJETO'},
-        {kind:'Class', description: 'CLASE'},  
-        {kind: 'Mixin', description: 'MIXIN'}
-    ]
+import { SelectKind } from "./SelectKind"
 
 export function NewEntityModal(props: {
     visible: boolean, 
@@ -21,9 +14,6 @@ export function NewEntityModal(props: {
     const [kind, setKind] = useState<Kind>('Singleton')
 
     
-    function toggleButtonColorByKind(aKind: Kind) {
-        return kind == aKind ? 'grey' : undefined
-    }
    
     return (
         <Portal>
@@ -33,20 +23,9 @@ export function NewEntityModal(props: {
                     onDismiss={() => props.setVisible(false)}
                 >
                     <TextInput placeholderTextColor="grey" onChangeText={setName} placeholder="Nombre definicion"></TextInput>
-                    <ToggleButton.Row style={{ marginVertical: 15, alignSelf: 'center'}}
-                        onValueChange={(value) => setKind(value as Kind)}
-                        value={kind}>
-                        {entityKinds.map(aKind => {
-                            return (
-                                <ToggleButton
-                                    key={aKind.kind}
-                                    style={{ width: 100, backgroundColor: toggleButtonColorByKind(aKind.kind) }}
-                                    icon={() => <View><Text>{aKind.description}</Text></View>}
-                                    value={aKind.kind}
-                                />
-                            )
-                        })}
-                    </ToggleButton.Row>
+                    
+                    <SelectKind kind={kind} setKind={setKind}/>
+
                     <Button
                         onPress={() => {
                             props.addEntity(fromJSON({ kind, name }))
